collision-handle: Clean up common functions
